+++ products/feature-flagging/feature-flagging-bootstrap/src/main/java/datadog/trace/api/featureflag/ufc/v1/ShardRange.java
@@ -1,10 +1,10 @@
 package datadog.trace.api.featureflag.ufc.v1;
 
 public class ShardRange {
-  public final int start;
-  public final int end;
+  public final long start;
+  public final long end;
 
-  public ShardRange(final int start, final int end) {
+  public ShardRange(final long start, final long end) {
     this.start = start;
     this.end = end;
   }
